Feature (reduce) Add accumulator for multiple options of same name
 - Add `Option::reduce` to take an accumulator function that will be
called on every token of the same name in the token list.

 - Add `Option::hasReducer` to check that an Option has a reducer
assigned.

 - Update parse function for the case where a named token has an
Option with a reducer, and the token has a value preceding it.
The accumulator with call a `map` if also assigned on the token
value before calling the accumulator with the previous accumulated
value and the token value.